Fix mypy errors and rename NamedTuple fields in gym.py
Add None to some type annotations in gym.py in order to fix some mypy
errors.

Rename some fields in the Gym NamedTuple class in order to clarify their
meanings and to fix a mypy error. The "count" field specifically causes
a mypy error because mypy interprets the field's name as the "count"
method that NamedTuple inherits from the base tuple class. According to
the mypy devs, this is expected behavior and is not a bug:
- https://github.com/python/mypy/issues/4507
- https://github.com/python/mypy/issues/17047